[TensorExpr] Implement shape inference for TE.
Since TE operates on a limited subset of ops with a well-defined
semantics, we can easily infer shapes of intermediate and output tensors
given shapes of the inputs.

There is a couple of ops that are not yet supported in the shape
inference, once we add them we could relax the shape info requirements
in the TE fuser: currently it requires all values in the fusion group to
have shapes known and we can change it to only inputs.

ghstack-source-id: efc2a46d58ff69d4efda53a3d3943bbbd65cd340
Pull Request resolved: https://github.com/pytorch/pytorch/pull/41451