[TensorExpr] Make tensorOrConstant, broadcast, chunk non-templatized functions.
Currently we used the template in order to be able to take both
`std::vector<ExprHandle>` and `std::vector<VarHandle>`. However,
semantics of this function tells that the only allowed option should be
the former one: we're specifying indices for the tensor access we want
to generate. While it could be convenient to avoid conversion from
vector of vars to a vector of exprs at the callsites, it makes the code
less explicit and thus more difficult to reason about.

ghstack-source-id: 64a07ce3276c6909ffcfae5595ab4204ff60f4d3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42202